Well then, a bit of a change...
OK, so this change is kind of big. Like way too big. But here's what I'm
thinking: Hooray has all the array methods shoved into a huge switch
statement; that's a maintenance nightmare. Also, it's not extensible.
What I really want to do is wrap up some of my application logic into
`links` that I can put into the `chain`. And I don't want to have to
define them in this library. But I'd also like some of the crazy taken
out of the PHP array methods.... so I /do/ want /some/ of the `links`
always available.

So, here's how I think it should work:

`Hooray` wraps an array and calls chainable methods. Oh, it can also
become the array that it wrapped via `realize`. That's about it.

`Chain` is a place where we keep all the links. By default, almost all
the PHP array methods should be wrapped in a chain and available. The
real fun comes when you want to make your own links.

When creating a new link you have to decide if it's terminal (returns
something other than an array, thusly breaking the chain), or if it
doesn't actually need the previous context (builder methods like fill
are like this). Then you have to return something, preferably something
that matches your intentions (array, scalar, null, whatevs).

Here are the boring details about actually using this thing:

`Chain::add` adds a new link to the chain. This link should return an
array, and when calling it you'll receive a `Hooray`.

`Chain::addTerminal` adds a new link to the chain as well. This link
should probably return something other than an array. You'll receive
whatever this method wanted to return when you call it, /not/ a
`Hooray`, so don't try to keep chaining when you run into a terminal
method.

`Chain::addWithoutContext` makes sure to remove the first argument
passed to itself when called. Don't worry, this first argument isn't
anything /you/ pass in - it's added to your args by `Hooray`, and it is
the array that a `Hooray` has wrapped.