Add a join method, and make sure Hooray unwraps
Joining. We all wish we could do it. Especially with such style as
`join`. This method accepts a joiner, an optional terminal
joiner (it will be used on the last element of the list), and also an
optional string to return if the array is empty. A convenience method
(`asEnglish`) prepopulates the joiner and terminal joiner for you.

You shouldn't have to keep track of whether a thing you have has already
been put into a hooray. Instead, let hooray do that for you. Go ahead,
try this: `hooray(hooray(hooray(1, 2, 3)))->realize()`.