More english, please
OK, `asEnglish` works better now. If given two items it'll use ` and `
instead of `, and ` as the terminal joiner. An even better solution may
be to have join always join the last element to the rest of the list
using the joiner _AND_ terminal joiner, but only if the `count > 2`.