Then first! Shuffle,
Shuffling is important to do sometimes. Hooray comes with a built-in
shuffle method, but you can of course override it if you don't like it.
The shuffling is done through a Fisher-Yates dance, and seems to be
deterministic, but also random.

First will return the first element of an array. If you have an
associative array, it will return the first element you added to the
array.